refactor: renomear arquivos e classes de autenticação para seguir convenções de nomenclatura Co-authored-by: Gabriel Bertolazi <gabriel.basto.bertolazi@gmail.com>
