+++ back/src/auth/auth.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Post, Body, InternalServerErrorException, UnauthorizedException } from '@nestjs/common';
-import { AuthService } from './auth.service'; // Ajuste conforme o caminho correto do serviço
+import { AuthService } from './auth.service'; 
 import { LoginDto } from './dto/login.dto';
 
 @Controller('auth')
@@ -8,6 +8,6 @@ export class AuthController {
 
   @Post('login')
   async login(@Body() loginDto: LoginDto) {
-    return await this.authService.login(loginDto); // Passe o objeto completo 'loginDto'
+    return await this.authService.login(loginDto); 
   }
 }
