+++ back/src/auth/auth.service.ts
@@ -1,13 +1,13 @@
 import { Injectable } from '@nestjs/common';
-import { JwtService } from '@nestjs/jwt'; // Importando o JwtService
+import { JwtService } from '@nestjs/jwt'; 
 import { LoginDto } from './dto/login.dto';
-import { PasswordAuth } from './PasswordAuth';
+import { PasswordAuth } from './password.auth';
 
 @Injectable()
 export class AuthService {
   constructor(
     private passwordAuth: PasswordAuth,
-    private tokenManager: JwtService, // Injetando o JwtService
+    private tokenManager: JwtService, 
   ) {}
 
   async login(loginDto: LoginDto) {
@@ -17,7 +17,6 @@ export class AuthService {
       throw new Error('Usuário ou senha inválidos');
     }
 
-    // Gerando o token JWT usando o método correto 'sign'
     const token = this.tokenManager.sign({
       id: user.id,
       email: user.email,
