+++ back/src/auth/password.auth.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@nestjs/common";
-import { PasswordAuthStrategy } from "./strategies/PasswordAuthStrategy";
-import { TokenManager } from "./tokenManager";
-import { UnauthorizedException } from "@nestjs/common"; // Importando a exceção
+import { PasswordAuthStrategy } from "./strategies/password.auth.strategy";
+import { TokenManager } from "./token.manager";
+import { UnauthorizedException } from "@nestjs/common"; 
 
 @Injectable()
 export class PasswordAuth {
@@ -11,15 +11,12 @@ export class PasswordAuth {
   ) {}
 
   async login(email: string, password: string): Promise<any> {
-    // Tenta autenticar o usuário
     const user = await this.strategy.authenticate(email, password);
     
-    // Se o usuário não for encontrado ou a autenticação falhar, lança a exceção 401
     if (!user) {
-      throw new UnauthorizedException('Invalid credentials'); // Lançando 401
+      throw new UnauthorizedException('Invalid credentials'); 
     }
     
-    // Se for válido, gera o token
     return this.tokenManager.generateToken({ id: user.id, email: user.email });
   }
 }
